Beta 0.11.8 (#27)
* Add prebuildify to the build

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Update prebuild.yml

* Fixes

* Update PackageNPMRelease.yml

* Update prebuildMacOSArm.yml

* Update prebuildMacOSArm.yml

* NPM Publish

* Update PackageNPMRelease.yml

* Update PackageNPMRelease.yml

* Update PackageNPMRelease.yml

* Update PackageNPMRelease.yml

* Update PackageNPMRelease.yml

* Update PackageNPMRelease.yml

* Update package.json

* prebuildify

* Remove darwin

* Create Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Remove unused github actions

* More logging

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Splitting the build, part 1

* Fix

* Simplify

* Simplify

* Update build.sh

* Update Dockerfile.oldDebian

* Update build.sh

* Update build.sh

* fixes

* Update build.sh

* Update build.sh

* Update Dockerfile.debian

* Try again

* Fix

* Update Prepare prebuild environment.yml

* Update build.sh

* Fixes

* Update build.sh

* Update build.sh

* Update Dockerfile.oldDebian

* Update Dockerfile.debian

* Update build.sh

* Try again

* Maybe

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Create Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Test Workflow.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Test Workflow.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Tweaking

* Tweaks

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Update Prepare prebuild environment.yml

* Fixes

* Cleanup

* Delete Test Workflow.yml
