--- .github/workflows/RemoveReleaseAssets.yml
@@ -1,36 +1,0 @@
-name: "Remove Release Assets from a GitHub Release"
-run-name: Remove Assets from ${{ github.event.inputs.tag }} - ${{ github.event.inputs.Bundle }}
-
-on:
-  workflow_dispatch:
-    inputs:
-      tag:
-        description: 'Tag'
-        required: true
-      Bundle:
-        description: 'Prebuild Bundle'
-        required: true
-        type: choice
-        options:
-        - darwin-arm64
-        - darwin-x64
-        - linux-arm
-        - linux-arm64tar.gz
-        - linux-ia32
-        - linux-x64
-        - linuxmusl-arm
-        - linuxmusl-arm64
-        - linuxmusl-x64
-        - win32-ia32
-        - win32-x64
-
-jobs:
-  Remove:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Delete old release assets
-        uses: mknejp/delete-release-assets@v1
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          tag: ${{ github.event.inputs.tag }}
-          assets: node-pty*${{ github.event.inputs.Bundle }}.tar.gz
