--- .github/workflows/prebuildMacOSArm.yml
@@ -1,40 +1,0 @@
-name: "Prebuild node-gyp MacOS ARM64 Binaries and package for a GitHub Release"
-run-name: Package MacOS ARM64 against ${{ github.event.inputs.tag }}
-
-on:
-  workflow_dispatch:
-    inputs:
-      tag:
-        description: 'Tag'
-        required: true
-
-jobs:
-  build:
-    runs-on: [self-hosted, macOS, ARM64]
-    steps:
-    - uses: actions/checkout@v4
-    - uses: actions/setup-node@v3
-      with:
-        node-version: lts/*
-
-    - name: macOS - Prebuild Binaries
-      if: runner.os == 'macOS'
-      run: |
-        npm ci --ignore-scripts
-        env JOBS=max node .prebuild/build.js
-
-    - name: Check prebuilds exist
-      shell: bash
-      run: |
-        fileCount=$(ls prebuilds/*/node-pty*tar.gz | wc -l)
-        echo "Prebuild count ${fileCount}"
-        echo "::notice::Prebuild count ${fileCount}, expected 25"
-        if [ "$fileCount" -eq 0 ]; then exit 1; fi
-        if [ "$fileCount" -ne 25 ]; then exit 1; fi
-
-    - uses: AButler/upload-release-assets@v2.0
-      name: Upload Release Artifacts to ${{ github.event.inputs.tag }}
-      with:
-        files: 'prebuilds/*/node-pty*tar.gz'
-        repo-token: ${{ secrets.GITHUB_TOKEN }}
-        release-tag: ${{ github.event.inputs.tag }}
