+++ .prebuild/Dockerfile.oldDebian
@@ -1,19 +1,27 @@
 ARG BASE_IMAGE
-FROM ${BASE_IMAGE:-library/debian:buster-slim}
+FROM ${BASE_IMAGE:-library/debian:9.6-slim}
 
 ARG QEMU_ARCH
 ENV QEMU_ARCH=${QEMU_ARCH:-x86_64}
 
 COPY .prebuild/qemu/qemu-${QEMU_ARCH}-static /usr/bin/
 
+RUN cat /etc/apt/sources.list
+RUN sed -i.bak 's/security.debian.org/archive.debian.org/g' /etc/apt/sources.list
+RUN sed -i.bak 's/deb.debian.org/archive.debian.org/g' /etc/apt/sources.list
+RUN sed -i.bak '/stretch-updates/d' /etc/apt/sources.list
 RUN apt-get update \
-  && apt-get install -y gcc g++ make python3 curl
+  && apt-get install -y gcc g++ make curl wget zlib1g-dev
 
 RUN [ "$(uname -m)" = "x86_64" ] && apt-get install -y g++-multilib gcc-multilib || echo "ok"
 
 ## Linux 32 bit support was dropped with NodeJS 18.4.0
 
-ENV NODE_VERSION 18.18.2
+RUN wget https://www.python.org/ftp/python/3.6.15/Python-3.6.15.tgz && tar xf Python-3.6.15.tgz \
+&& cd Python-3.6.15 && ./configure --prefix=/usr --enable-optimizations --enable-shared && make -j8 build_all \
+&& make -j8 altinstall && update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.6 50
+
+ENV NODE_VERSION 16.20.2
 
 RUN case "${QEMU_ARCH}" in \
     x86_64) NODE_ARCH='x64';; \
@@ -29,6 +37,8 @@ RUN case "${QEMU_ARCH}" in \
   && rm "node-v$NODE_VERSION-linux-$NODE_ARCH.tar.gz" \
   && ln -s /usr/local/bin/node /usr/local/bin/nodejs \
   && npm set prefix /usr/local
+
+
 
 RUN mkdir /node-pty
 WORKDIR /node-pty
