+++ .prebuild/build.sh
@@ -5,7 +5,7 @@
 cd /node-pty
 # rm -rf node_modules
 
-echo "Building for $(uname -m)..."
+echo "Building for $(uname -m)... ->" $*
 
 npm ci --ignore-scripts
 # https://github.com/microsoft/vscode/blob/c23f0305dbf82b2319b198f4dbf3c5d5bc522f15/build/azure-pipelines/linux/product-build-linux-client.yml#L113-L125
@@ -15,5 +15,11 @@ npm ci --ignore-scripts
  fi
 
 #node .prebuild/build.js
-env JOBS=max node .prebuild/buildify.js
-env JOBS=max node .prebuild/build.js
+env JOBS=max node $*
+
+#env JOBS=max node .prebuild/prebuild.js -t 19.0.0 -t 20.0.0 -t 21.0.0
+#env JOBS=max node .prebuild/prebuildify.js -t 19.0.0 -t 20.0.0 -t 21.0.0
+
+#env JOBS=max node .prebuild/prebuild.js -t 10.0.0 -t 11.0.0 -t 12.0.0 -t 13.0.0 -t 14.0.0 -t 15.0.0 -t 16.0.0 -t 17.0.1 -t 18.0.0
+#env JOBS=max node .prebuild/prebuildify.js -t 10.0.0 -t 11.0.0 -t 12.0.0 -t 13.0.0 -t 14.0.0 -t 15.0.0 -t 16.0.0 -t 17.0.1 -t 18.0.0
+#env JOBS=max node .prebuild/electron.js -t 5.0.0 -t 6.0.0 -t 7.0.0 -t 8.0.0 -t 9.0.0 -t 10.0.0 -t 11.0.0 -t 12.0.0 -t 13.0.0 -t 14.0.0 -t 15.0.0 -t 16.0.0 -t 17.0.0 -t 18.0.0
